refactor(mimir): address review — split PagesView, remove inline styles, use Table (NIU-668)
- Split 600-line PagesView into TreeNode, ZoneBlock, ZoneRenderers, MetaPanel components
- Remove all 41 inline styles; use CSS custom properties as data channels for dynamic values
  (--mm-tree-depth for indent, --mm-glyph-size for PageTypeGlyph font-size)
- Replace hand-rolled table in SourcesView with @niuulabs/ui Table component
- Fix initial page-selection bug: useState initialises before data loads → useEffect
- Extract magic numbers to named constants (ZONE_SAVE_RESET_DELAY_MS, FEED_LIMIT, etc.)
- Add error state rendering to OverviewView
- All 870 tests passing
