+++ web-next/packages/plugin-mimir/src/ui/MimirPage.css
@@ -7,15 +7,18 @@
   display: flex;
   align-items: center;
   gap: var(--space-3);
-  margin-bottom: var(--space-4);
+  border-bottom: 1px solid var(--color-border-subtle);
+  padding: var(--space-4) var(--space-6);
 }
 
 .mimir-page__header h2 {
   margin: 0;
+  font-size: var(--text-lg);
 }
 
 .mimir-page__subtitle {
   color: var(--color-text-secondary);
+  margin: 0;
 }
 
 .mimir-page__status {
